+++ src/request.rs
@@ -24,7 +24,6 @@ pub struct SentRequest<Body> {
 pub struct ReceivedRequest<Body> {
 	request_id: u32,
 	service_id: i32,
-	body: Body,
 	incoming_rx: mpsc::UnboundedReceiver<Message<Body>>,
 	command_tx: mpsc::UnboundedSender<Command<Body>>,
 }
@@ -32,7 +31,7 @@ pub struct ReceivedRequest<Body> {
 /// An incoming request or stream message.
 pub enum Incoming<Body> {
 	/// An incoming request.
-	Request(ReceivedRequest<Body>),
+	Request(ReceivedRequest<Body>, Body),
 
 	/// An incoming stream message.
 	Stream(Message<Body>),
@@ -150,14 +149,12 @@ impl<Body> ReceivedRequest<Body> {
 	pub(crate) fn new(
 		request_id: u32,
 		service_id: i32,
-		body: Body,
 		incoming_rx: mpsc::UnboundedReceiver<Message<Body>>,
 		command_tx: mpsc::UnboundedSender<Command<Body>>,
 	) -> Self {
 		Self {
 			request_id,
 			service_id,
-			body,
 			incoming_rx,
 			command_tx,
 		}
@@ -171,11 +168,6 @@ impl<Body> ReceivedRequest<Body> {
 	/// Get the service ID of the initial request message.
 	pub fn service_id(&self) -> i32 {
 		self.service_id
-	}
-
-	/// Get the body of the initial request message.
-	pub fn body(&self) -> &Body {
-		&self.body
 	}
 
 	/// Receive the next update message of the request from the remote peer.
@@ -237,17 +229,8 @@ impl<Body> std::fmt::Debug for ReceivedRequest<Body> {
 impl<Body> std::fmt::Debug for Incoming<Body> {
 	fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
 		match self {
-			Self::Request(x) => {
-				write!(f, "ReceivedRequest(")?;
-				x.fmt(f)?;
-				write!(f, ")")?;
-			},
-			Self::Stream(x) => {
-				write!(f, "Stream(")?;
-				x.fmt(f)?;
-				write!(f, ")")?;
-			},
+			Self::Request(x, _body) => write!(f, "Request({:?})", x),
+			Self::Stream(x) => write!(f, "Stream({:?})", x),
 		}
-		Ok(())
 	}
 }