+++ src/request_tracker.rs
@@ -102,7 +102,7 @@ impl<Body> RequestTracker<Body> {
 		request_id: u32,
 		service_id: i32,
 		body: Body,
-	) -> Result<ReceivedRequest<Body>, error::DuplicateRequestId> {
+	) -> Result<(ReceivedRequest<Body>, Body), error::DuplicateRequestId> {
 		match self.received_requests.entry(request_id) {
 			Entry::Occupied(_entry) => {
 				// TODO: Check if the channel is closed so we don't error out unneccesarily.
@@ -123,7 +123,7 @@ impl<Body> RequestTracker<Body> {
 			Entry::Vacant(entry) => {
 				let (incoming_tx, incoming_rx) = mpsc::unbounded_channel();
 				entry.insert(incoming_tx);
-				Ok(ReceivedRequest::new(request_id, service_id, body, incoming_rx, self.command_tx.clone()))
+				Ok((ReceivedRequest::new(request_id, service_id, incoming_rx, self.command_tx.clone()), body))
 			},
 		}
 	}
@@ -148,8 +148,8 @@ impl<Body> RequestTracker<Body> {
 	pub async fn process_incoming_message(&mut self, message: Message<Body>) -> Result<Option<Incoming<Body>>, ProcessIncomingMessageError> {
 		match message.header.message_type {
 			MessageType::Request => {
-				let received_request = self.register_received_request(message.header.request_id, message.header.service_id, message.body)?;
-				Ok(Some(Incoming::Request(received_request)))
+				let (received_request, body) = self.register_received_request(message.header.request_id, message.header.service_id, message.body)?;
+				Ok(Some(Incoming::Request(received_request, body)))
 			},
 			MessageType::Response => {
 				self.process_incoming_response(message).await?;
@@ -254,7 +254,7 @@ mod test {
 		});
 
 		// Simulate an incoming request and an update.
-		let_assert!(Ok(Some(Incoming::Request(mut received_request))) = tracker.process_incoming_message(Message::request(1, 2, Body)).await);
+		let_assert!(Ok(Some(Incoming::Request(mut received_request, _body))) = tracker.process_incoming_message(Message::request(1, 2, Body)).await);
 		assert!(let Ok(None) = tracker.process_incoming_message(Message::requester_update(1, 10, Body)).await);
 
 		// Receive the update.